KohaQuick: partially implement getting items checked out today
This is for implementing the Print Checkouts tab, using the REST API. Much of the work is done in the commit, but it's not complete: I stopped when I learned that I would not be getting REST API access to the production Koha. I will disable some of this code in a future commit.
Version 0.4.6.0.